Add files via upload
Cambio en el tipo de dato de todo lo calculado.
Se pasa de int a float todo excepto el factorial.

+++ TP_1_Cascara/GalloAlejandro.h
@@ -3,41 +3,41 @@
 /**************************************************************************************************/
 
 
-/** \brief  Suma 2 enteros
+/** \brief  Suma 2 float
  *
  * \param Primer Operando a Sumar
  * \param Segundo Operando a Sumar
  * \return Retorna la suma de los parametros.
  *
  */
-int sumar(int n1, int n2);
+float sumar(float n1, float n2);
 
-/** \brief  Resta 2 enteros
+/** \brief  Resta 2 floats
  *
  * \param Primer Operando a restar
  * \param Segundo Operando a restar
  * \return Retorna la resta de los parametros
  *
  */
-int restar(int n1, int n2);
+float restar(float n1, float n2);
 
-/** \brief  Divide el Primer entero por el segundo. Retorna 0 y un mensaje de error si el Divisor es 0.
+/** \brief  Divide el Primer float por el segundo. Retorna 0 y un mensaje de error si el Divisor es 0.
  *
  * \param Primer Operando - Dividendo
  * \param Segundo Operando  - Divisor
  * \return Retorna la division (float), entre el Primer Operando y el Segundo
  *
  */
-float dividir(int nro1, int nro2);
+float dividir(float nro1, float nro2);
 
-/** \brief  Multiplica 2 Enteros
+/** \brief  Multiplica 2 floats
  *
  * \param Primer Operando a multiplicar
  * \param Segundo Operando a multiplicar
  * \return Devuelve la multiplicacion entre el Primer y segundo operando
  *
  */
-int multiplicar(int nro1, int nro2);
+float multiplicar(float nro1, float nro2);
 
 /** \brief Devuelve el Factorial de un entero.
  *
@@ -47,18 +47,12 @@ int multiplicar(int nro1, int nro2);
  */
 int factorial(int nro);
 
-/** \brief Obtiene un entero por teclado.
+/** \brief Obtiene un float por teclado.
  *
  * \return Retorna el nro obtenido por teclado.
  *
  */
-int obtenerNro();
+float obtenerNro();
 
-/** \brief  devuelve el mayor int de un array.
- *
- * \param Array a analizar
- * \param Tamaño del array
- * \return Retorne el mayor entero.
- *
- */
+
